Move iamge_upload controller logic to new concern
