+++ app/controllers/application_controller.rb
@@ -5,10 +5,9 @@ class ApplicationController < ActionController::Base
   include ValidRequest
   include Pundit
   include FastlyHeaders
+  include ImageUploads
 
   rescue_from ActionView::MissingTemplate, with: :routing_error
-
-  MAX_FILENAME_LENGTH = 250
 
   def not_found
     raise ActiveRecord::RecordNotFound, "Not Found"
@@ -78,31 +77,5 @@ class ApplicationController < ActionController::Base
 
   def touch_current_user
     current_user.touch
-  end
-
-  def log_image_data_to_datadog
-    images = Array.wrap(
-      params.dig("user", "profile_image") ||
-      params.dig("podcast", "image") ||
-      params.dig("organization", "profile_image") ||
-      params["image"],
-    )
-
-    raise if images.empty?
-
-    images.each do |image|
-      tags = [
-        "controller:#{params['controller']}",
-        "action:#{params['action']}",
-        "content_type:#{image.content_type}",
-        "original_filename:#{image.original_filename}",
-        "tempfile:#{image.tempfile}",
-        "size:#{image.size}",
-      ]
-
-      DatadogStatsClient.increment("image_upload_error", tags: tags)
-    end
-
-    raise
   end
 end
