+++ app/controllers/concerns/image_uploads.rb
@@ -1,0 +1,31 @@
+module ImageUploads
+  extend ActiveSupport::Concern
+
+  MAX_FILENAME_LENGTH = 250
+
+  def log_image_data_to_datadog
+    images = Array.wrap(
+      params.dig("user", "profile_image") ||
+      params.dig("podcast", "image") ||
+      params.dig("organization", "profile_image") ||
+      params["image"],
+    )
+
+    raise if images.empty?
+
+    images.each do |image|
+      tags = [
+        "controller:#{params['controller']}",
+        "action:#{params['action']}",
+        "content_type:#{image.content_type}",
+        "original_filename:#{image.original_filename}",
+        "tempfile:#{image.tempfile}",
+        "size:#{image.size}",
+      ]
+
+      DatadogStatsClient.increment("image_upload_error", tags: tags)
+    end
+
+    raise
+  end
+end
